+++ src/AngleSharp.Wasm/IWasmCompiledModule.cs
@@ -1,0 +1,10 @@
+namespace AngleSharp.Wasm;
+
+using System;
+
+/// <summary>
+/// Represents a compiled WebAssembly module.
+/// </summary>
+public interface IWasmCompiledModule : IDisposable
+{
+}