+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.0
+
+Released on Friday, July 31 2026.
+
+- Added `wasmtime` as Wasm engine
+- Added integration with exposure as DOM objects
+
 # 0.13.0
 
 Released on Sunday, June 6 2021.